refactor(Annotator): extract document content save into shared hook
Both AnnotatorFindReplaceModal and AnnotatorSearchLine duplicated the
document update mutation. Move it into useDocumentContentSave so the
save logic and its onSuccess/onError toasts live in one place.
